requirements_airflow.txt
mojito2
psycopg2-binary
pandas
requests